+++ front/components/actions/dust_app_run/DustAppRunActionDetails.tsx
@@ -5,9 +5,9 @@ import { useMemo } from "react";
 
 import { ActionDetailsWrapper } from "@app/components/actions/ActionDetailsWrapper";
 import type { ActionDetailsComponentBaseProps } from "@app/components/actions/types";
+import { CodeBlock } from "@app/components/assistant/markdown/CodeBlock";
 import type { GetContentToDownloadFunction } from "@app/components/assistant/markdown/ContentBlockWrapper";
 import { ContentBlockWrapper } from "@app/components/assistant/markdown/ContentBlockWrapper";
-import { CodeBlock } from "@app/components/markdown/CodeBlock";
 
 export function DustAppRunActionDetails({
   action,
